backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.5.0
transformers==4.40.2
huggingface-hub==0.23.0
python-dotenv==1.0.1
